Avoid `bytestring` >= 0.11 with GHC < 8.0
In `bytestring` 0.11 the `PS` constructor becomes a pattern synonym, only
available with GHC >= 8.0.  Since this package depends on `bytestring`
internals, and in particular the `PS` constructor, with GHC < 8.0 it must
stick to older bytestring releases.

With GHC >= 8.0, this package will be compatible with the imminent
`bytestring` 0.11, so the dependency ceiling is for now set to < 0.12.