Handle sign at chunk boundary
When the leading plus or minus sign of the (potential) integer is a
chunk boundary, we need to "peek" at the next chunk to determine
whether it in fact starts with a digit, before continuing an unsigned
number with a starting value of zero.

Now that we're peeking, we may end up discharging a monadic effect,
which must not be repeated, so on failure (the result is not a digit),
we must return a new 'Chunk' with the sign as prefix.  We must not
return the original chunk containing the sign, because its continuation
could contain an effect we've already processed.

Updated the tests to check that readInt evalutes each effect in the
stream exactly once, and added some tests to verify that leading signs
at chunk boundaries work as expected.