Use faster unsafeWithForeignPtr for GHC 9.0+
With GHC 9.0 `withForeignPtr` was made safe for actions that might loop forever
or always throw an exception.  This safety comes at the cost of performance,
the result is always allocated, which is not ideal for primitive byte-level
memory operations.

A new `unsafeWithForeignPtr` is added that provides the old non-allocating
behaviour when the action is sure to terminate.  With GHC 9.0 and later,
we use this function to loop over the content of ByteStream chunks.